success page orderDetails and fixes

+++ app/api/cart/route.ts
@@ -6,6 +6,14 @@ export async function DELETE(req: Request) {
   const { userId } = auth();
 
   try {
+    const cart = await prisma.cart.findUnique({
+      where: {
+        userId: userId!,
+      },
+    });
+
+    if (!cart) return NextResponse.json("No cart to delete");
+
     await prisma.cart.delete({
       where: {
         userId: userId!,
